+++ src/main/java/racingcar/Application.java
@@ -5,6 +5,7 @@ import racingcar.controller.RacingCarController;
 import java.util.Scanner;
 
 public class Application {
+
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
         RacingCarController racingCarController = new RacingCarController(scanner);
