+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,7 @@
 package racingcar.domain;
 
 public class Car {
+
     private String name;
     private int position;
