+++ src/main/java/racingcar/view/RacingCarError.java
@@ -1,6 +1,7 @@
 package racingcar.view;
 
 public class RacingCarError {
+
     private static final String OVERLAPPED_MESSAGE = "[ERROR] 이름을 중복하여 사용할 수 없습니다";
     private static final String LOWER_LENGTH_MESSAGE = "[ERROR] 이름은 1자 이상으로 입력해주세요";
     private static final String UPPER_LENGTH_MESSAGE = "[ERROR] 이름은 5자 이하로 입력해주세요";
