+++ src/main/java/racingcar/view/RacingCarView.java
@@ -3,6 +3,7 @@ package racingcar.view;
 import racingcar.domain.Car;
 
 public class RacingCarView {
+
     private static final String INPUT_MESSAGE = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n";
     private static final String TURN_INPUT_MESSAGE = "시도할 회수는 몇회인가요?\n";
     private static final String RESULT_INDICATION_MESSAGE = "실행 결과\n";
