+++ src/test/java/racingcar/utils/RandomNumberGeneratorTest.java
@@ -13,9 +13,8 @@ public class RandomNumberGeneratorTest {
 
     @Test
     void randomNumberGenerator() {
-        for(int i = FIRST_TRYING; i <LAST_TRYING ; i++) {
+        for (int i = FIRST_TRYING; i < LAST_TRYING; i++) {
             int value = RandomNumberGenerator.turnNumber();
-
             assertTrue(value <= UP_LIMIT);
             assertTrue(value >= DOWN_LIMIT);
         }
