+++ src/test/java/racingcar/view/RacingCarViewTest.java
@@ -10,6 +10,7 @@ import java.io.PrintStream;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class RacingCarViewTest {
+
     private static final String INPUT_MESSAGE = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n";
     private static final String TURN_INPUT_MESSAGE = "시도할 회수는 몇회인가요?\n";
     private static final String SAMPLE_CAR_NAME = "sample";
@@ -49,6 +50,6 @@ public class RacingCarViewTest {
     void showWinner() {
         String sampleWinners = SAMPLE_WINNERS;
         RacingCarView.showWinner(sampleWinners);
-        assertEquals(SAMPLE_WINNERS+WIN_MESSAGE, outputStreamCaptor.toString());
+        assertEquals(SAMPLE_WINNERS + WIN_MESSAGE, outputStreamCaptor.toString());
     }
 }
